Move Workspace Permission types from model to backend-spi
-  This was really just dangling here without any reason

JIRA: RAIL-1815
