+++ libs/sdk-backend-spi/src/workspace/permissions/index.ts
@@ -1,5 +1,35 @@
 // (C) 2019-2020 GoodData Corporation
-import { IWorkspacePermissions } from "@gooddata/sdk-model";
+
+/**
+ * Workspace permission types
+ *
+ * @public
+ */
+export type WorkspacePermission =
+    | "canInitData"
+    | "canUploadNonProductionCSV"
+    | "canExecuteRaw"
+    | "canExportReport"
+    | "canAccessWorkbench"
+    | "canCreateReport"
+    | "canCreateVisualization"
+    | "canCreateAnalyticalDashboard"
+    | "canManageMetric"
+    | "canManageReport"
+    | "canManageAnalyticalDashboard"
+    | "canManageProject"
+    | "canCreateScheduledMail"
+    | "canListUsersInProject"
+    | "canManageDomain";
+
+/**
+ * Dictionary of workspace permissions
+ *
+ * @public
+ */
+export type IWorkspacePermissions = {
+    [permission in WorkspacePermission]: boolean;
+};
 
 /**
  * Service to query workspace permissions
